onboarding/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
stripe==7.8.0
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
email-validator==2.1.0
pytz==2023.3.post1
Werkzeug==3.0.1
gunicorn==21.2.0
bcrypt==4.1.2
PyJWT==2.8.0
celery==5.3.4
pydantic==2.5.3
prometheus-flask-exporter==0.23.0